+++ src/Enqueue/Blocks/EnqueueBlocks.php
@@ -263,11 +263,10 @@ class EnqueueBlocks extends AbstractEnqueueBlocks
 		$output = $this->getEnqueueSharedInlineCommonItems();
 
 		// Frontend part.
-		$hideGlobalMessageTimeout = HooksHelpers::getFilterName(['block', 'form', 'hideGlobalMsgTimeout']);
 		$redirectionTimeout = HooksHelpers::getFilterName(['block', 'form', 'redirectionTimeout']);
 		$fileRemoveLabel = HooksHelpers::getFilterName(['block', 'file', 'previewRemoveLabel']);
 
-		$output['hideGlobalMessageTimeout'] = \apply_filters($hideGlobalMessageTimeout, 6000);
+		$output['hideGlobalMessageTimeout'] = (int) SettingsHelpers::getOptionValueWithFallback(SettingsSettings::SETTINGS_GENERAL_HIDE_GLOBAL_MSG_TIMEOUT, '6') * 1000;
 		$output['redirectionTimeout'] = \apply_filters($redirectionTimeout, 300);
 		$output['labels'] = [
 			'selectOptionAria' => \esc_html__('Select option', 'eightshift-forms'),